+++ oap-server/server-core/src/main/java/org/apache/skywalking/oap/server/core/CoreModuleConfig.java
@@ -135,14 +135,13 @@ public class CoreModuleConfig extends ModuleConfig {
     private String searchableLogsTags = "";
 
     /**
-     * The number of threads used to synchronously refresh the metrics data to the storage. When the value is negative,
-     * the number of processors will used as the default value.
+     * The number of threads used to synchronously refresh the metrics data to the storage. When the value is zero or
+     * negative, the number of processors will used as the default value.
      *
      * @since 8.5.0
      */
-    @Getter
     @Setter
-    private int syncThreads = 10;
+    private int syncThreads = -1;
 
     /**
      * The maximum number of processes supported for each synchronous storage operation. When the number of the flush
@@ -154,6 +153,10 @@ public class CoreModuleConfig extends ModuleConfig {
 
     public CoreModuleConfig() {
         this.downsampling = new ArrayList<>();
+    }
+
+    public int getSyncThreads() {
+        return this.syncThreads > 0 ? this.syncThreads : Runtime.getRuntime().availableProcessors();
     }
 
     /**